Moved ProxyManagerTest base class to net.lightbody.bmp.proxy.test.util to parallel browsermob-core package changes
